requirements.txt
streamlit
pandas
plotly
snowflake-connector-python==3.12.0
python-dotenv